lidar_trend_design_app_v3/requirements.txt
fastapi
uvicorn
numpy
scipy
shapely
open3d>=0.18.0
laspy
pydantic
pillow
torch
diffusers
transformers
accelerate
opencv-python
feedparser
sentence-transformers
scikit-learn